% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddGitHubIssue.R
\name{AddGitHubIssue}
\alias{AddGitHubIssue}
\title{AddGitHubIssue}
\usage{
AddGitHubIssue(endpoint = "POST /repos/DrMattG/InterEco/issues", title, body)
}
\arguments{
\item{endpoint}{The repo where you want the issue to be written (default is InterEco)}

\item{title}{Title of the issue}

\item{body}{The description of the issue}
}
\value{
Adds GitHub issue to the package repository
}
\description{
AddGitHubIssue
}
\examples{
\dontrun{
AddGitHubIssue(title="Test issue", body="Test automatic issue")
}
}
