% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsuni.R
\docType{data}
\name{antsuni}
\alias{antsuni}
\title{antsuni}
\format{
A data frame with 121 rows and 46 variables:
\describe{
\item{\code{index}}{double COLUMN_DESCRIPTION}
\item{\code{SITE_ID}}{character COLUMN_DESCRIPTION}
\item{\code{Abundance}}{double COLUMN_DESCRIPTION}
\item{\code{Richness}}{double COLUMN_DESCRIPTION}
\item{\code{Evenness}}{double COLUMN_DESCRIPTION}
\item{\code{Genus_Richness}}{double COLUMN_DESCRIPTION}
\item{\code{X}}{double COLUMN_DESCRIPTION}
\item{\code{Y}}{double COLUMN_DESCRIPTION}
\item{\code{Elev}}{double COLUMN_DESCRIPTION}
\item{\code{Farm_ID}}{character COLUMN_DESCRIPTION}
\item{\code{c10grp}}{double COLUMN_DESCRIPTION}
\item{\code{c12grp}}{double COLUMN_DESCRIPTION}
\item{\code{c15grp}}{double COLUMN_DESCRIPTION}
\item{\code{Max_temp}}{double COLUMN_DESCRIPTION}
\item{\code{Ltrain}}{double COLUMN_DESCRIPTION}
\item{\code{LT_clim}}{double COLUMN_DESCRIPTION}
\item{\code{R_clim}}{double COLUMN_DESCRIPTION}
\item{\code{R_weath}}{double COLUMN_DESCRIPTION}
\item{\code{Woody}}{double COLUMN_DESCRIPTION}
\item{\code{GrndCvr}}{double COLUMN_DESCRIPTION}
\item{\code{Litter}}{double COLUMN_DESCRIPTION}
\item{\code{Logs}}{double COLUMN_DESCRIPTION}
\item{\code{LUI}}{double COLUMN_DESCRIPTION}
\item{\code{Exotic_grnd}}{double COLUMN_DESCRIPTION}
\item{\code{Bare}}{double COLUMN_DESCRIPTION}
\item{\code{perc.C}}{double COLUMN_DESCRIPTION}
\item{\code{C_N}}{double COLUMN_DESCRIPTION}
\item{\code{Bulk_density}}{double COLUMN_DESCRIPTION}
\item{\code{perc.Clay}}{double COLUMN_DESCRIPTION}
\item{\code{EC}}{double COLUMN_DESCRIPTION}
\item{\code{Irid_ab}}{double COLUMN_DESCRIPTION}
\item{\code{Rhyt_ab}}{double COLUMN_DESCRIPTION}
\item{\code{Pheid_ab}}{double COLUMN_DESCRIPTION}
\item{\code{Mono_ab}}{double COLUMN_DESCRIPTION}
\item{\code{Nylan_ab}}{double COLUMN_DESCRIPTION}
\item{\code{Melop_ab}}{double COLUMN_DESCRIPTION}
\item{\code{Card_ab}}{double COLUMN_DESCRIPTION}
\item{\code{Camp_rich}}{double COLUMN_DESCRIPTION}
\item{\code{Irid_rich}}{double COLUMN_DESCRIPTION}
\item{\code{Melop_rich}}{double COLUMN_DESCRIPTION}
\item{\code{Meran_rich}}{double COLUMN_DESCRIPTION}
\item{\code{Mono_rich}}{double COLUMN_DESCRIPTION}
\item{\code{Pheid_rich}}{double COLUMN_DESCRIPTION}
\item{\code{Rhyt_rich}}{double COLUMN_DESCRIPTION}
\item{\code{Stig_rich}}{double COLUMN_DESCRIPTION}
\item{\code{Tetra_rich}}{double COLUMN_DESCRIPTION}
}
}
\usage{
antsuni
}
\description{
Ant biodiversity sampled along a 1000 m elevation gradient in eastern Australia (as a proxy for a climate gradient) from sites with variable land cover and land use.
}
\details{
DETAILS
}
\keyword{datasets}
