% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestmod.R
\docType{data}
\name{bestmod}
\alias{bestmod}
\title{internal_model}
\format{
A glm model object
}
\usage{
bestmod
}
\description{
A glm model of the \link[InterEco]{ants_data}
}
\details{
Call:  glmer(richness~scale(lt_clim)*scale(woody) +scale(lt_clim)*scale(perc_clay)+  scale(lt_clim)*scale(bare)+ scale(lui) +scale(exotic_grnd)  + (1|farm_id), family="poisson", data=dat)
}
\keyword{datasets}
