% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_function.R
\name{custom_function}
\alias{custom_function}
\title{custom_function for plotting}
\usage{
custom_function(data, mapping, method = "loess", ...)
}
\arguments{
\item{data}{data to be plotted}

\item{mapping}{mapping of aes}

\item{method}{method to be used defaults to loess}

\item{...}{additional params}
}
\value{
smoothed loess line
}
\description{
custom_function for plotting
}
