% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_obj.R
\docType{data}
\name{fit_obj}
\alias{fit_obj}
\title{fit_obj}
\format{
A lm model object and a data frame
}
\usage{
fit_obj
}
\description{
A lm model and data object (combined). For testing
}
\details{
data and model just for testing
}
\keyword{datasets}
