% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\docType{data}
\name{input}
\alias{input}
\title{input}
\format{
A glm model object and a data frame
}
\usage{
input
}
\description{
A glm model and data object (combined). For testing
}
\details{
data and model just for testing
}
\keyword{datasets}
