% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_model.R
\docType{data}
\name{internal_model}
\alias{internal_model}
\title{internal_model}
\format{
A glmer model object
}
\usage{
internal_model
}
\description{
Bestmodel for the antsuni data.  Generalized linear mixed model fit by maximum likelihood (Laplace Approximation) \link{'glmerMod'}
}
\details{
DETAILS
}
\keyword{datasets}
