% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\docType{data}
\name{mod}
\alias{mod}
\title{internal_model}
\format{
A glm model object
}
\usage{
mod
}
\description{
A glm model of the \link[InterEco]{fake_data}
}
\details{
Call:  glm(formula = abun ~ forest \* elev \* taxon, family = poisson, data = dat)
}
\keyword{datasets}
