% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerunModel.R
\name{rerunModel}
\alias{rerunModel}
\title{Rerun the model for use with some plot types}
\usage{
rerunModel(input)
}
\arguments{
\item{input}{model and data list object}
}
\value{
rerun model object
}
\description{
Rerun the model for use with some plot types
}
