% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runInterEco.R
\name{runInterEco}
\alias{runInterEco}
\title{runInterEco
Run the shinyApp associated with interEco}
\usage{
runInterEco()
}
\value{
launches the interEco ShinyApp
}
\description{
runInterEco
Run the shinyApp associated with interEco
}
